+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.7
+
+- Rename package to `arbxcel_with_options`
+- Add name string value trimming
+- Add translation key auto-converted to camel case (for flutter_gen compatibility)
+- Change default sheet name to `Sheet1` - for google sheet export compatibility
+- Add option to change sheet name when parsing .xlsx file
+- Add `@@locale` key with locale e.g. `en` as value for 1st line of translated .arb files
+- Change type of inserted string value as `Object` instead of `String`
+
 ## 0.0.6
 
 - Rename bin name to `arbxcel`